+++ src/modules/obsOperators_mod.f90
@@ -1090,13 +1090,13 @@ contains
        end do
 
        if (Vcode == 5002) then
-          ! case with top thermo level above top momentum level (Vcode=5002)
+          ! case with top thermo level above top momentum level
           do jl = 1, nwndlev
              zuu(jl) = col_getElem(columnTrlOnTrlLev,jl  ,headerIndex,'UU')
              zvv(jl) = col_getElem(columnTrlOnTrlLev,jl  ,headerIndex,'VV')
           end do
-       elseif (Vcode == 5005 .or. Vcode == 5100 .or. Vcode == 21001) then
-          ! case without top thermo above top momentum level or unstaggered (Vcode=5001/4/5)
+       elseif (Vcode == 5005 .or. Vcode == 21001) then
+          ! case without top thermo above top momentum level or unstaggered
           do jl = 1, nwndlev-1
              zuu(jl) = col_getElem(columnTrlOnTrlLev,jl+1,headerIndex,'UU')
              zvv(jl) = col_getElem(columnTrlOnTrlLev,jl+1,headerIndex,'VV')
@@ -1104,7 +1104,6 @@ contains
           zuu(nwndlev) = zuu(nwndlev-1)
           zvv(nwndlev) = zuu(nwndlev-1)
        else
-          ! case with Vcode /= 5002 and Vcode /= 5005 and Vcode /= 5100
           call utl_abort('oop_gpsro_nl: invalid vertical coord!')
        end if
        zuu(ngpslev) = zuu(nwndlev)